gateway_service/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.28.0
PyJWT>=2.6.0
cryptography>=39.0.0
redis>=4.5.1
channels-redis>=4.0.0
channels>=4.0.0
websockets>=10.4
uvicorn>=0.27.0
psycopg2-binary>=2.9.5